requirements.txt
Django
djangorestframework
nltk
numpy
pandas
PyMySQL
rake_nltk
requests
scikit_learn
spacy
textstat
